fix: what several review passes found in the refactor
Three defects, two of them mine and both able to cost a user everything
in config.json:

- withoutNulls recursed into objects but not arrays, so a null in
  excludedApps, textLanguages, a zone set's rects or a workspace's items
  still threw, and a throw is the whole file set aside. It now walks
  lists too, which is what its doc comment already claimed.
- migrated() ran before withoutNulls and asked `stored["workspaces"] ==
  nil`, which is false for an explicit null. A pre-workspaces config with
  "workspaces": null skipped the layouts migration and lost every saved
  layout, silently, with the file reported as readable. Nulls are now
  stripped first, so the question means what it says.
- applyConfig reassigned updates.automatic on every write and neither
  didSet behind it compared values, so UpdateSchedule.start() rebuilt the
  24-hour timer from zero each time. A Mac whose settings are touched
  once a day would never have checked for an update again.

Both decoding defects have tests that fail without the fix.

Behaviour the wider fan-out had started trampling:

- MouseTools.apply stops only a tap that is running. stop() also hides
  the overlay, and a find-the-cursor flash has no tap behind it, so an
  unrelated write used to blink it out under the user's hand.
- AppearanceSettings.apply checks before assigning. Setting NSApp
  .appearance makes every open window re-resolve and redraw, and
  dragging the zone colour writes config per tick.
- clamp() no longer filters textLanguages. What Vision recognises is a
  fact about this Mac, not a bound on the setting, and clamping on load
  meant a language chosen on one Mac was deleted from a synced file by
  the first save on another. The picker only offers what is supported
  and TextExtractor names an unsupported tag back to the caller.

The rest is the branch finally keeping its own rules. AGENTS.md said
every manager owns an apply(_ config:) and that nothing pushes a single
field; only two of eight did. DragSnapManager, GrabMove, MouseTools,
HotkeyManager, NewWindowWatcher and UpdateManager now have one, and
applyConfig is a list of apply() calls. The accessibility-grant handler
calls applyConfig instead of hand-repeating three of them. Config owns
what its own modifier names mean, and ZoneAppearance takes a Config, so
the overlay and the pointer tools cannot end up different colours.

Also: ten refreshWorkspaceModel/refreshZoneModel calls that didMutate
already covers, RectangleImport's duplicate clamp, two bounds ZonesTuning
spelled out instead of reading from Config, an orphaned MARK, three
imports whose symbols moved out of AppDelegate, and five file headers
citing a line limit their file is no longer over.

UpdateManager 333 -> 152 and GrabMove 349 -> 312, both split along the
seam the added apply() pushed them past; UpdateManager leaves the
baseline. CHANGELOG covers the two user-visible changes: clamping on
load, and a null no longer costing the whole file.
