+++ scripts/line-limit-baseline
@@ -10,7 +10,6 @@
 # lines  path
 489   App/Tests/plonkTests/ConfigTests.swift
 444   App/Tests/plonkTests/RouterTests.swift
-349   App/Sources/plonk/GrabMove.swift
-333   App/Sources/plonk/UpdateManager.swift
+312   App/Sources/plonk/GrabMove.swift
 308   App/Sources/plonk/WorkspaceLauncher.swift
 306   mcp/src/api.ts
